fix between validation and controller

+++ src/user/controller.js
@@ -1,7 +1,6 @@
 const User = require('./model');
 const bcrypt = require('bcryptjs');
 const jwt = require('jsonwebtoken');
-const validator = require('validator');
 const ObjectID = require('mongodb').ObjectId;
 
 const SECRET_KEY = process.env.SECRET_KEY;
@@ -13,6 +12,10 @@ const register = async (req, res, next) => {
         }
         try {
             const email = req.body.email;
+            const user = await User.findOne({ email });
+            if (user) {
+                return res.status(400).json({ message: ['Email already exist!'] });
+            }
             const password = hashedPassword;
             const refreshToken = '';
             const name = 'Admin';
@@ -31,6 +34,19 @@ const login = async (req, res, next) => {
         const email = req.body.email;
         const password = req.body.password;
         const user = await User.findOne({ email });
+        let errors = [];
+        if (!user) {
+            errors.push('User not found!');
+        }
+        else {
+            const validPassword = await bcrypt.compare(password, user.password);
+            if (!validPassword) {
+                errors.push('Password does not matched!');
+            }
+        }
+        if (!errors.length == 0) {
+            return res.status(400).json({ message: errors });
+        }
         const token = generateToken(user);
         const refreshToken = jwt.sign({ userID: user._id }, SECRET_KEY);
         const updatedUser = await User.updateOne(
@@ -49,18 +65,33 @@ const login = async (req, res, next) => {
 }
 
 const newToken = async (req, res, next) => {
-    try{
+    try {
         const refreshToken = req.body.refreshToken;
         const userID = req.body.userID;
-        const user = await User.findOne({_id: ObjectID(userID)});
+        let errors = [];
+
+        const user = await User.findOne({ _id: ObjectID(userID) });
+        if (!user) {
+            errors.push('User not found!');
+        }
+        else {
+            if (refreshToken != user.refreshToken) {
+                errors.push('Invalid refresh token!');
+            }
+        }
+
+        if (!errors.length == 0) {
+            return res.status(400).json({ message: errors });
+        }
+
         const isValidRefreshToken = jwt.verify(refreshToken, SECRET_KEY);
-        if(isValidRefreshToken){
+        if (isValidRefreshToken) {
             const token = generateToken(user);
             res.json({ token });
         }
     }
-    catch(err){
-        res.status(400).json({message: [err]})
+    catch (err) {
+        res.status(400).json({ message: [err] })
     }
 }
 
@@ -69,51 +100,80 @@ const generateToken = (user) => {
 }
 
 const resetPassword = async (req, res, next) => {
-    const userID = req.body.userID;
-    const newPassword = req.body.newPassword;
-    try{
+    try { 
+        const userID = req.body.userID;
+        const newPassword = req.body.newPassword;
+
+        const user = await User.findOne({ _id: ObjectID(userID) });
+        if (!user) {
+            return res.status(400).json({ message: ['User not found!'] });
+        }
+
         const hashedPassword = await bcrypt.hash(newPassword, 10);
-        if(hashedPassword){
+        if (hashedPassword) {
             const updatedUser = await User.updateOne(
-                        { _id: ObjectID(userID) },
-                        {
-                            $set: {
-                                password: hashedPassword
-                            }
-                        }
-                    );
+                { _id: ObjectID(userID) },
+                {
+                    $set: {
+                        password: hashedPassword
+                    }
+                }
+            );
             res.json({ message: 'Password successfully changed' });
         }
     }
-    catch(err){
-        res.status(400).json({message: [err]})
+    catch (err) {
+        res.status(400).json({ message: [err] })
     }
 }
 
 const getUsers = async (req, res, next) => {
-    try{
+    try {
         const users = await User.find();
         res.json({ users });
     }
-    catch(err){
+    catch (err) {
         res.status(400).json({ message: [err] });
     }
 }
 
 const getUserByID = async (req, res, next) => {
-    try{
+    try {
         const userID = req.params.userID;
-        const user = await User.findOne({_id: ObjectID(userID)});
+        const user = await User.findOne({ _id: ObjectID(userID) });
+        if (!user) {
+            return res.status(400).json({ message: ['User not found!'] });
+        }
         res.json({ user });
     }
-    catch(err){
+    catch (err) {
         res.status(400).json({ message: [err] });
     }
 }
 
 const updateUserByID = async (req, res, next) => {
-    try{
+    try {
         const user = req.body.user;
+        let errors = [];
+        const isUserValid = await User.findOne({ _id: ObjectID(user._id) });
+        if (!isUserValid) {
+            errors.push('User not found!');
+        }
+        else {
+            const validPassword = await bcrypt.compare(user.password, isUserValid.password);
+            if (!validPassword) {
+                errors.push('Password does not matched!');
+            }
+            const isUserEmail = await User.findOne({ email: user.email });
+            if (isUserEmail) {
+                if (isUserValid.email !== isUserEmail.email) {
+                    errors.push('Email already exist!');
+                }
+            }
+        }
+        if (!errors.length == 0) {
+            return res.status(400).json({ message: errors });
+        }
         const updatedUser = await User.updateOne(
             { _id: ObjectID(user._id) },
             {
@@ -125,29 +185,32 @@ const updateUserByID = async (req, res, next) => {
         )
         res.json({ message: 'User successfully updated!' });
     }
-    catch(err){
+    catch (err) {
         res.status(400).json({ message: [err] });
     }
 }
 
 const deleteUserByID = async (req, res, next) => {
-    try{
+    try {
         const userID = req.params.userID;
         const user = await User.findOne({ _id: ObjectID(userID) });
-        const deletedUser = await User.deleteOne({ _id: ObjectID(userID)});
+        if (!user) {
+            return res.status(400).json({ message: ['User not found!'] });
+        }
+        const deletedUser = await User.deleteOne({ _id: ObjectID(userID) });
         res.json({ message: 'User successfully deleted!' });
     }
-    catch(err){
-        res.status(400).json({ message: [err] });   
+    catch (err) {
+        res.status(400).json({ message: [err] });
     }
 }
 
 const deleteUsers = async (req, res, next) => {
-    try{
+    try {
         await User.deleteMany();
         res.json({ message: 'Users successfully deleted' });
     }
-    catch(err){
+    catch (err) {
         res.status(400).json({ message: [err] });
     }
 }
